Consume ESP32 changes for SDK uplift. (#473)
Mirror the changes made in the ESP32 repository in this repository.

+++ esp32/components/F7Services/System/FileSystem.cpp
@@ -314,7 +314,7 @@ void FileSystem::WriteFile(Message *message)
             std::string s = PARTITION_PATH_NAME;
             s += "/";
             s += fileData->Name;
-            TRACE_MESSAGE("Writing %d bytes into file %s", fileData->ContentsLength, s.c_str());
+            TRACE_MESSAGE("Writing %u bytes into file %s", (unsigned int) fileData->ContentsLength, s.c_str());
             TRACE_HEX_BUFFER(fileData->Contents, fileData->ContentsLength);
             FILE *file = fopen(s.c_str(), "w");
             if (file != NULL)
